app works with most scratch work i did trimmed off
