+++ package/broom.py
@@ -4,15 +4,62 @@ from collections import namedtuple
 from typing import List
 from itertools import combinations
 from functools import reduce
-import threading
 import time
-import simple_app
-from .simple_app import playround
 
-thread = threading.Thread(target=simple_app.playround)
-thread.start()
-result_available = threading.Event()
-result = None
+
+
+def returnJSON(game_object):
+  deck = {
+    "card_store": {},
+    "deck_order": []
+  }
+
+  for k, v in game_object.deck.card_store.items():
+    deck['card_store'][k] = v
+
+  # deck['deck_order'] = list(game_object.deck.card_store.keys())
+  deck['deck_order'] = game_object.deck.order()
+
+  pl1 = {
+    "score": '',
+    "hand": [],
+    "name": ''
+  }
+
+  pl1['score'] = game_object.pl1.score
+  pl1['name'] = game_object.pl1.name
+  pl1['hand'] = game_object.pl1.get_hand()
+
+  pl2 = {
+    "score": '',
+    "hand": [],
+    "name": ''
+  }
+
+  pl2['score'] = game_object.pl2.score
+  pl2['name'] = game_object.pl2.name
+  pl2['hand'] = game_object.pl2.get_hand()
+
+  table_cards = []
+
+  table_cards = game_object.get_table_cards()
+
+  # if len(game_object.table_cards) > 0:
+  #   for x in range(len(game_object.table_cards)):
+  #     table_cards.append(x)
+
+  return {
+    "game": {
+      "pl1": pl1,
+      "pl2": pl2,
+      "deck": deck,
+      "table_cards": table_cards
+    }
+  }
+  
+def playsJSON(plays_set):
+  return
+
 
 Card = namedtuple("Card", ['suit', 'face', 'owner'])
 faces = list(range(1,11))
@@ -68,12 +115,52 @@ class Deck:
   def order(self):
     return self.deck_order
 
+class Play:
+  def __init__(self, Card={}):
+    self.card = Card
+  def get_card(self):
+    return self.card
+
+class PlayList:
+  def __init__(self, owner, plays=set()):
+    self.owner = owner
+    self.plays = plays
+
+  def get_plays(self):
+    return list(self.plays)
+  def get_owner(self):
+    return self.owner
+  def add_play(self, Play):
+    self.plays.add(Play)
+    return self.plays
+
+# print(CardStore)
+# playlist = []
+# pl1_pl = PlayList('Player1')
+# for key, card in CardStore.items():
+#   #playlist.append(Play(card))
+#   pl1_pl.add_play(Play(card))
+
+
+
+# #pl = PlayList(playlist).get_plays()
+# pl = pl1_pl.get_plays()
+
+# for card in pl:
+#   print(card.card)
+# #print(pl.plays.card)
+
+
 class Player: 
   score = 0
   hand = set()
   name = ''
-  def __init__(self, name, hand=[]):
+  def __init__(self, name, score, hand=[], actual_play=[], ):
     self.name = name
+    # added score and hand to init so they can be passed in to deserializer
+    self.score = score
+    self.hand = hand if hand is not None else []
+    self.playlist = PlayList(name)
   def play_turn(self,deck):
     return NotImplemented
   def new_hand(self, cards):
@@ -94,19 +181,51 @@ class Player:
       if len(good_hands) > 0:
         return random.choice(good_hands)
       return random.choice(list(play))
-    return play.pop()
+    # return player object for frontend   
+    self.actual_play = play.pop()  
+    return self
+    # original game logic to return just one card
+    #return play.pop
+  def get_hand(self):
+    return list(self.hand)
+  #deserialzer  
+  @classmethod
+  def from_json(cls, data):
+      return cls(**data)
+
 
 class Game:
   deck = Deck()
-  pl1 = Player('p1')
-  pl2 = Player('p2')
+  pl1 = Player('p1', 0)
+  pl2 = Player('p2', 0)
   table_cards = set()
-  def __init__(self, pl1, pl2, deck=Deck()):
+  paused = True
+
+  def __init__(self, pl1, pl2, deck=Deck(), valid_plays=[]):
     self.pl1 = pl1
     self.pl2 = pl2
     self.deck = deck
+    self.valid_plays = valid_plays if valid_plays is not None else []
+
     print("Start game")
     #return NotImplemented
+
+   # serialzer method
+  def toJSON(self):
+    # return json.loads(json.dumps(self, default=lambda o: o.__dict__ if type(o.__dict__) == '__dict__' else o.__set__, 
+    #   sort_keys=True))
+    return json.loads(json.dumps(self, default=lambda o: o.__dict__ , sort_keys=True))
+    
+
+  def set_pause_state(self, switch):
+    print(f"#### switching paused to:{switch} ####")
+    self.paused = switch
+
+  def get_table_cards(self):
+    return list(self.table_cards)
+
+  def get_pause_state(self):
+    return self.paused
 
   def set_card_owner(self, card, owner):
     c = self.deck.cards()[card]
@@ -118,12 +237,12 @@ class Game:
   def deal_hand(self):
     p1 = set()
     p2 = set()
-    print(len(self.deck.order()))
+    print(f"cards_remaining_before_deal: {len(self.deck.order())}")
     # deal out to p1 and p2 alternating 3 each
     for count in range(0,3):
       [ p1.add(d) for d in self.deck.deal()]
       [ p2.add(d) for d in self.deck.deal()]
-    print(len(self.deck.order())) 
+    print(f"cards_remaining_after_deal: {len(self.deck.order())}")
     return p1,p2
 
   def deal_start(self):
@@ -142,7 +261,8 @@ class Game:
           plays.add(combo)
       else:
         plays.add(tuple(player.hand))
-    return plays
+    self.valid_plays_list = list(plays)
+    return self
 
   def apply_play(self,play, player):
     # validate(play)
@@ -217,11 +337,21 @@ class Game:
       self.pl2.award_point()
     print(f'Points:\tPL1\tPL2\nOros:\t[{len(p1_oros)}]\t[{len(p2_oros)}]\nSevens:\t[{len(p1_sevens)}]\t[{len(p2_sevens)}]\nCards:\t[{len(p1_total)}]\t[{len(p2_total)}]')
 
-  def play_round(self, first_player, second_player):
+
+
+  def play_first_round(self, first_player, second_player):
     p1_cards, p2_cards ,table_cards = self.deal_start()
     first_player.new_hand(p1_cards)
     second_player.new_hand(p2_cards)
     self.table_cards = table_cards
+    return self
+
+  def get_best_play(self, player):
+    playable = self.valid_plays(player,self.table_cards)
+    player.get_play(playable)
+    return self 
+
+  def play_round(self, first_player, second_player):
     
     last_scored = ''
     cards_left = len(self.deck.order())
@@ -232,6 +362,9 @@ class Game:
         second_player.new_hand(p2_cards)
         cards_left = len(self.deck.order())
 
+      def get_play(play):
+        #print(play)
+        return play
 
       # hand per player
       while (len(first_player.hand) + len(second_player.hand) > 0):
@@ -244,10 +377,10 @@ class Game:
           playable = self.valid_plays(second_player,self.table_cards)
           play = second_player.get_play(playable)
           if self.apply_play(play,second_player): last_scored = second_player.name
-        while not result_available.wait(timeout=120):
-          print('waiting for user input')
-        print(result)
-        print(play)
+        while (self.paused):
+          #print(first_player)
+          get_play(play)
+          
     # award last_player_to_score remaining cards
     [self.set_card_owner(card_id, last_scored) for card_id, card in self.deck.cards().items() if card.owner == '']
     self.apply_score()
@@ -255,20 +388,5 @@ class Game:
   def print_score(self):
         print("Player 1 score: {}\nPlayer 2 score: {}".format(self.pl1.score, self.pl2.score))
     
-
-if __name__ == "__main__":
-  
-  p1 = Player('player_1')
-  p2 = Player('player_2')
-  deck = Deck(make_deck())
-  g = Game(p1,p2,deck)
-  
-  rounds = 0
-  while (p1.score < 15 and p2.score < 15):
-    rounds += 1
-    g.reset_deck()
-    if (rounds % 2 == 1):
-      g.play_round(p1,p2)
-    else:
-      g.play_round(p2,p1)
-    print("Round {}:\n\tPlayer 1 score: {}\n\tPlayer 2 score: {}".format(rounds, p1.score, p2.score))
+def test():
+    return 'test'
