--- package/test.py
@@ -1,8 +1,0 @@
-import sys; import pprint
-pprint.pprint(sys.path)
-
-# sys.path.append("C:\\Users\\soria\\Repos\\poet-escoba\\package")
-pprint.pprint(sys.path)
-
-# from package import package
-# print(package)
