--- sql/Untitled-1
@@ -1,52 +1,0 @@
-('Hearts', '1',  '2'),
-('Hearts', '1',  '3'),
-('Hearts', '1',  '4'),
-('Hearts', '1',  '5'),
-('Hearts', '1',  '6'),
-('Hearts', '1',  '7'),
-('Hearts', '1',  '8'),
-('Hearts', '1',  '9'),
-('Hearts', '1',  '10'),
-('Hearts', '1',  'Jack'),
-('Hearts', '1',  'Queen'),
-('Hearts', '1',  'King'),
-('Hearts', '1',  'Ace'),
-('Clubs', '1',  '2'),
-('Clubs', '1',  '3'),
-('Clubs', '1',  '4'),
-('Clubs', '1',  '5'),
-('Clubs', '1',  '6'),
-('Clubs', '1',  '7'),
-('Clubs', '1',  '8'),
-('Clubs', '1',  '9'),
-('Clubs', '1',  '10'),
-('Clubs', '1',  'Jack'),
-('Clubs', '1',  'Queen'),
-('Clubs', '1',  'King'),
-('Clubs', '1',  'Ace'),
-('Spades', '1',  '2'),
-('Spades', '1',  '3'),
-('Spades', '1',  '4'),
-('Spades', '1',  '5'),
-('Spades', '1',  '6'),
-('Spades', '1',  '7'),
-('Spades', '1',  '8'),
-('Spades', '1',  '9'),
-('Spades', '1',  '10'),
-('Spades', '1',  'Jack'),
-('Spades', '1',  'Queen'),
-('Spades', '1',  'King'),
-('Spades', '1',  'Ace'),
-('Diamonds', '1',  '2'),
-('Diamonds', '1',  '3'),
-('Diamonds', '1',  '4'),
-('Diamonds', '1',  '5'),
-('Diamonds', '1',  '6'),
-('Diamonds', '1',  '7'),
-('Diamonds', '1',  '8'),
-('Diamonds', '1',  '9'),
-('Diamonds', '1',  '10'),
-('Diamonds', '1',  'Jack'),
-('Diamonds', '1',  'Queen'),
-('Diamonds', '1',  'King'),
-('Diamonds', '1',  'Ace');
